// src/api/uploads/enums/asset-type.enum.ts
export enum AssetType {
  AVATAR = 'AVATAR',
  VERIFICATION_IMAGE = 'VERIFICATION_IMAGE',
  VERIFICATION_SCREENSHOT = 'VERIFICATION_SCREENSHOT',
  VERIFICATION_AUDIO = 'VERIFICATION_AUDIO',
  VERIFICATION_VIDEO = 'VERIFICATION_VIDEO',
  LESSON_THUMBNAIL = 'LESSON_THUMBNAIL',
  COURSE_THUMBNAIL = 'COURSE_THUMBNAIL',
  CHALLENGE_MEDIA = 'CHALLENGE_MEDIA',
  BADGE_ICON = 'BADGE_ICON',
  REPORT_ASSET = 'REPORT_ASSET',
}